+++ danceschool/core/views.py
@@ -2,7 +2,7 @@ from django.http import HttpResponseRedirect, Http404, HttpResponseBadRequest
 from django.shortcuts import get_object_or_404, get_list_or_404
 from django.conf import settings
 from django.core.exceptions import ObjectDoesNotExist
-from django.core.urlresolvers import reverse
+from django.urls import reverse
 from django.views.generic import FormView, CreateView, UpdateView, DetailView, TemplateView, ListView
 from django.db.models import Min, Q, Count
 from django.utils.translation import ugettext_lazy as _
@@ -33,7 +33,7 @@ from .mixins import (EmailRecipientMixin, StaffMemberObjectMixin, FinancialConte
                      AdminSuccessURLMixin, EventOrderMixin, SiteHistoryMixin)
 from .signals import get_customer_data
 from .utils.requests import getIntFromGet
-from .utils.timezone import ensure_localtime, ensure_timezone
+from .utils.timezone import ensure_timezone
 
 
 # Define logger for this file
@@ -968,7 +968,7 @@ class RepeatEventsView(SuccessMessageMixin, AdminSuccessURLMixin, PermissionRequ
             # For each new occurrence, we determine the new startime by the distance from
             # midnight of the first occurrence date, where the first occurrence date is
             # replaced by the date given in repeat list
-            old_min_time = ensure_localtime(event.startTime).replace(hour=0, minute=0, second=0, microsecond=0)
+            old_min_time = event.localStartTime.replace(hour=0, minute=0, second=0, microsecond=0)
 
             old_occurrence_data = [
                 (x.startTime - old_min_time, x.endTime - old_min_time, x.cancelled)